JDP220502-14: Encja Cart - zmieniony typ int na long w cartId i totalValue

+++ src/main/java/com/kodilla/ecommercee/domain/Cart.java
@@ -19,11 +19,11 @@ public class Cart {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     @Column(name = "Cart_ID", unique = true)
-    private int cartId;
+    private long cartId;
 
     @Column(name = "Total_Value")
     @NotNull
-    private int totalValue;
+    private long totalValue;
 
     /*
     @ManyToOne
