+++ src/main/java/com/kodilla/ecommercee/domain/Order.java
@@ -21,7 +21,7 @@ public class Order {
     @NotNull
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     @Column(name = "Order_ID")
-    private int OrderId;
+    private long OrderId;
 
 
     @OneToOne
